Updated bone connection script. Was wrongly connecing also "twist" bones. Added support to Align Mixamo animations taken from A pose.
